updated project

+++ src/utils/numbers_utils.py
@@ -4,7 +4,7 @@ NEAREST_TEN = -1
 NEAREST_HUNDRED = -2
 
 
-def is_between_inclusive(value, lower: int, upper: int) -> bool:
+def is_between_inclusive(value, lower: float, upper: float) -> bool:
     assertions.not_none(value)
     assertions.not_none(lower, value_name='lower')
     assertions.not_none(upper, value_name='upper')
